[visionOS] REGRESSION(299348@main): Timeline on inline video controls doesn't glow anymore, making it hard to interact with https://bugs.webkit.org/show_bug.cgi?id=301196 <rdar://161134512>
Reviewed by NOBODY (OOPS!).

Go back to setting the TimeControl's children at layout time in order
for the VisionSlider customization to be picked up.

Add support for flexbox layouts to the VisionSlider.

Test: media/modern-media-controls/visionos-inline-media-controls/visionos-inline-media-controls-layout.html

* LayoutTests/TestExpectations:
* LayoutTests/media/modern-media-controls/visionos-inline-media-controls/visionos-inline-media-controls-layout-expected.txt: Added.
* LayoutTests/media/modern-media-controls/visionos-inline-media-controls/visionos-inline-media-controls-layout.html: Added.
* LayoutTests/platform/visionos/TestExpectations:
Add a basic visionOS specific test to cover the regression.

* Source/WebCore/Modules/modern-media-controls/controls/time-control.js:
(TimeControl.):
Don't set the children from the constructor.
(TimeControl.prototype._performIdealLayout):
Set children during layout like we did before 299348@main.

(TimeControl.prototype.this.remainingTimeLabel): Deleted.
* Source/WebCore/Modules/modern-media-controls/controls/vision-slider.js:
(VisionSlider.prototype.commit):
When part of ta flexbox layout, the width is not set on the Slider.
Add support for this setup using a percentage based width for the
primary fill.

* Source/WebCore/Modules/modern-media-controls/js-files:
Make the `VisionMediaControls` testable.